+++ lib/draper/railtie.rb
@@ -51,7 +51,8 @@ module Draper
 
     console do
       require 'action_controller/test_case'
-      ApplicationController.new.view_context
+      controller = ApplicationController.new
+      controller.view_context if controller.respond_to? :view_context
       Draper::ViewContext.build
     end